UI Framework Initial Commit
This is a pretty good start on the UI code, so I'm backing it up as a
precautionary measure. Some of the code is quite messy, and it is
largely untested as I am too tired right now to actually bother to
write the makefile for this thing and add the setup code to
UIApplication.cpp, but it should theoretically work fine. I will test
it and commit again when I wake up.